Testing/requirements-gemini.txt
streamlit
python-dotenv
genai
google-generativeai
fitz
pypdf
